It seems to work with the abstract classes now
The problem might have been that the admin_init was called inside an admin_menu action and/or that the hooks were not called early enough in the chain.

Though the latter should not have been a problem because initialize_components in the index.php should automatically have caused the base_class' construct to add those actions.