Still works but it is too un-organized
I fixed some of the errors shown in WordPress, but the structure and logic are too scrambled and un-organized. The base classes need to be optimized and restructured again, the current structure makes it difficult to include sub-menus.

Am going to restructure the base classes separated by logic and function, this removes the need for separate pages/sections, just a singular new page for each menu item.