Testing adding new submenu
I think the directory structure needs to change.
It is smarter to make a directory for each menu item, this way everything that is related stays together.

The whole naming structure needs to change as well, things need to be named to clearly imply its function/logic.

I need to figure out the naming, directory structure  and separation of code.

All (display)text related to each menu should be separated and merged into a single location (array). This way the output/render html can also be separated and be programmatically inserted, heredoc might be the best option to do this.