Added `populate_allergen_table()`
Populates the `wp_allergens_dietary_ictoria_allergy` table with the `Allergens_Dietary_Ictoria_Functions::default_options()`. It instantiates on load, there does need to be a check added to prevent the populate function to run if it has already run.

Now I can add the html that needs to be output.