Changed page `render_html`
Removed the fields related code as the input fields come from the forms now. The `render_html` now loops through the sections' `section_callback`, this way each section can easily be wrapped in divs or whatever is needed.

Now the html being output is also more customizable because we are no longer stuck in the WordPress structure.

Do note that the pages still follow WordPress styling, this is because of the `<div class="wrap"></div>` that is around each page.