Compatibility for PRO version (#93)
* made a start on cleaning the free from the pro
Need edit the tables for quick edits.
Tabs for the CRUD functionality of allergens
DB to check what part is used for the pro and what for the free version

* See message on the free version
The before mentioned of files that need to be done is still an estimate.
When everything has changed vissibility change the UML and leave a note on the pro
That it extends on free and that these extended methods will not be noted unless it has some other different feature to be added

As tip update the last number of the version to keep track on what you have done
Unfortunately I did not add comments since I forgot to add them in experimenting on things

* safety commit

* working quick edit

only the change status is not working now

* Classes updated

I have now fixed the problems i was having before and succesfully extended every class from the free to the pro classes.

* separation of edit and add

* Edit page and other fixes

* Working edit and edit link from table

* Working messages for update page

Also made functions for the add allergen for readability.

* Delete fixed and change allergen

I fixed delete query/function and made changes to the cookie of change allergen.

* Requested changes + update page

* removed comments

* Update for multiple attachments

* Did some fixing

* fixed compat for forms
Add/edit allergen
Delete single allergen (on table)

The following still needs to be fixxed:
Bulk Delete (on table)

The following still needs to be tested if it works:
Update allergen(bulk/singular image update)

* bulk delete
Fixed the bulk delete on custom allergens
Encountered and fixxed a bug where default allergens where also deleted

Next up is the bulk update form
First to test and if there is something wrong bug fix

* UML
Much needed changes made to the UML
Working from top to bottom in the UML to check every single file

For next time start at the DB package

* got rid of extra return

* UML
edited the rest of the class diagram

---------

Co-authored-by: Verdi-B <verdi.burgers@ictoria.nl>
Co-authored-by: Jake <Jake.van.de.kolk@ictoria.nl>